requirements.txt
# Зависимости для Bitrix24 бота

# Web framework
Flask==3.0.0

# HTTP requests
requests==2.31.0

# Существующие зависимости из основного проекта
pandas==2.1.4
reportlab==4.0.7
Pillow==10.1.0
openpyxl==3.1.2
python-barcode==0.15.1